#53 Extract tests into own files according extracted core.Dogu types
