+++ core/dogu_v2_test.go
@@ -1,9 +1,6 @@
 package core
 
 import (
-	"fmt"
-	"slices"
-	"sort"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -43,21 +40,6 @@ func TestGetEnvironmentVariablesAsStringArray(t *testing.T) {
 		EnvironmentVariables: []EnvironmentVariable{sampleEnv},
 	}
 	assert.Equal(t, []string{"HAPPY_MODE=true"}, dogu.GetEnvironmentVariablesAsStringArray())
-}
-
-func TestDependsOn(t *testing.T) {
-	dogu := Dogu{
-		Name: "hansolo",
-		Dependencies: []Dependency{
-			{Type: DependencyTypeDogu, Name: "a"},
-			{Type: DependencyTypeDogu, Name: "b"},
-		},
-		OptionalDependencies: []Dependency{{Type: DependencyTypeDogu, Name: "c"}},
-	}
-	assert.True(t, dogu.DependsOn("a"))
-	assert.True(t, dogu.DependsOn("b"))
-	assert.False(t, dogu.DependsOn("c"))
-	assert.False(t, dogu.DependsOn("d"))
 }
 
 func TestGetVersion(t *testing.T) {
@@ -129,14 +111,6 @@ func TestDoguIsNewerThan(t *testing.T) {
 
 	newer, _ = dogu.IsNewerThan(other)
 	assert.False(t, newer)
-}
-
-func TestUnmarshalProperties(t *testing.T) {
-	dogu := &Dogu{}
-	dogu, _, err := ReadDoguFromFile("../resources/test/unmarshalProperties.json")
-	require.Nil(t, err)
-	assert.Equal(t, "http://test.test", dogu.Properties["logoutUrl"])
-	assert.Equal(t, "25", dogu.Properties["TestPort"])
 }
 
 func TestHasExposedCommand_shouldFindCommand(t *testing.T) {
@@ -222,134 +196,6 @@ func Test_getNamespace(t *testing.T) {
 	assert.Equal(t, "official", actual)
 }
 
-func TestDogu_Dependencies(t *testing.T) {
-	// given
-	depDogu1 := Dependency{Type: DependencyTypeDogu, Name: "DoguTest"}
-	depDogu2 := Dependency{Type: DependencyTypeDogu, Name: "DoguTest2"}
-	depClient1 := Dependency{Type: DependencyTypeClient, Name: "ClientTest1"}
-	depClient2 := Dependency{Type: DependencyTypeClient, Name: "ClientTest2"}
-	depPackage1 := Dependency{Type: DependencyTypePackage, Name: "PackageTest1"}
-	depPackage2 := Dependency{Type: DependencyTypePackage, Name: "PackageTest2"}
-
-	t.Run("GetAllDependenciesOfType - Return only empty slice when dependencies are empty", func(t *testing.T) {
-		// given
-		dogu1 := &Dogu{
-			Name: "namespace/dogu1",
-		}
-
-		// when
-		dependenciesDogu := dogu1.GetAllDependenciesOfType(DependencyTypeDogu)
-
-		// then
-		assert.Empty(t, dependenciesDogu)
-	})
-
-	t.Run("GetDependenciesOfType - Return only empty slice when dependencies are empty", func(t *testing.T) {
-		// given
-		dogu1 := &Dogu{
-			Name: "namespace/dogu1",
-		}
-
-		// when
-		dependenciesDogu := dogu1.GetDependenciesOfType(DependencyTypeDogu)
-
-		// then
-		assert.Empty(t, dependenciesDogu)
-	})
-
-	t.Run("GetOptionalDependenciesOfType - Return only empty slice when dependencies are empty", func(t *testing.T) {
-		// given
-		dogu1 := &Dogu{
-			Name: "namespace/dogu1",
-		}
-
-		// when
-		dependenciesDogu := dogu1.GetOptionalDependenciesOfType(DependencyTypeDogu)
-
-		// then
-		assert.Empty(t, dependenciesDogu)
-	})
-
-	t.Run("GetAllDependenciesOfType - Return all dependencies with their correct types", func(t *testing.T) {
-		// given
-		dogu1 := &Dogu{
-			Name:                 "namespace/dogu1",
-			Dependencies:         []Dependency{depDogu1, depClient1, depPackage1},
-			OptionalDependencies: []Dependency{depDogu2, depClient2, depPackage2},
-		}
-
-		// when
-		dependenciesDogu := dogu1.GetAllDependenciesOfType(DependencyTypeDogu)
-		dependenciesClient := dogu1.GetAllDependenciesOfType(DependencyTypeClient)
-		dependenciesPackage := dogu1.GetAllDependenciesOfType(DependencyTypePackage)
-
-		// then
-		assert.NotEmpty(t, dependenciesDogu)
-		assert.Len(t, dependenciesDogu, 2)
-		assert.Equal(t, depDogu1, dependenciesDogu[0])
-		assert.Equal(t, depDogu2, dependenciesDogu[1])
-		assert.Len(t, dependenciesClient, 2)
-		assert.Equal(t, depClient1, dependenciesClient[0])
-		assert.Equal(t, depClient2, dependenciesClient[1])
-		assert.Len(t, dependenciesPackage, 2)
-		assert.Equal(t, depPackage1, dependenciesPackage[0])
-		assert.Equal(t, depPackage2, dependenciesPackage[1])
-	})
-
-	t.Run("GetDependenciesOfType - Return only required dependencies with their correct types", func(t *testing.T) {
-		// given
-		dogu1 := &Dogu{
-			Name:                 "namespace/dogu1",
-			Dependencies:         []Dependency{depDogu1, depClient1, depPackage1},
-			OptionalDependencies: []Dependency{depDogu2, depClient2, depPackage2},
-		}
-
-		// when
-		dependenciesDogu := dogu1.GetDependenciesOfType(DependencyTypeDogu)
-		dependenciesClient := dogu1.GetDependenciesOfType(DependencyTypeClient)
-		dependenciesPackage := dogu1.GetDependenciesOfType(DependencyTypePackage)
-
-		// then
-		assert.Len(t, dependenciesDogu, 1)
-		assert.Equal(t, depDogu1, dependenciesDogu[0])
-		assert.Len(t, dependenciesClient, 1)
-		assert.Equal(t, depClient1, dependenciesClient[0])
-		assert.Len(t, dependenciesPackage, 1)
-		assert.Equal(t, depPackage1, dependenciesPackage[0])
-	})
-
-	t.Run("GetOptionalDependenciesOfType - Return all optional dependencies with their correct types", func(t *testing.T) {
-		// given
-		dogu1 := &Dogu{
-			Name:                 "namespace/dogu1",
-			Dependencies:         []Dependency{depDogu1, depClient1, depPackage1},
-			OptionalDependencies: []Dependency{depDogu2, depClient2, depPackage2},
-		}
-
-		// when
-		dependenciesDogu := dogu1.GetOptionalDependenciesOfType(DependencyTypeDogu)
-		dependenciesClient := dogu1.GetOptionalDependenciesOfType(DependencyTypeClient)
-		dependenciesPackage := dogu1.GetOptionalDependenciesOfType(DependencyTypePackage)
-
-		// then
-		assert.Len(t, dependenciesDogu, 1)
-		assert.Equal(t, depDogu2, dependenciesDogu[0])
-		assert.Len(t, dependenciesClient, 1)
-		assert.Equal(t, depClient2, dependenciesClient[0])
-		assert.Len(t, dependenciesPackage, 1)
-		assert.Equal(t, depPackage2, dependenciesPackage[0])
-	})
-}
-
-func Test_sortDogus(t *testing.T) {
-	unsortedDogus := []*Dogu{{Name: "Dogu1", Version: "11.22.33-1"}, {Name: "Dogu2", Version: "0.1.3-5"}, {Name: "Dogu3", Version: "0.5.3-3"}, {Name: "Dogu4", Version: "9.3.9"}}
-	expectedDogus := []*Dogu{{Name: "Dogu1", Version: "11.22.33-1"}, {Name: "Dogu4", Version: "9.3.9"}, {Name: "Dogu3", Version: "0.5.3-3"}, {Name: "Dogu2", Version: "0.1.3-5"}}
-
-	sort.Sort(ByDoguVersion(unsortedDogus))
-
-	assert.Equal(t, unsortedDogus, expectedDogus)
-}
-
 func TestContainsDogu(t *testing.T) {
 	dogua := createTestDogu(t, "dogua", "1.0.0-1")
 	doguav2 := createTestDogu(t, "dogua", "1.0.0-2")
@@ -372,202 +218,4 @@ func TestContainsDogu(t *testing.T) {
 func createTestDogu(t *testing.T, name string, version string) *Dogu {
 	t.Helper()
 	return &Dogu{Name: "official/" + name, Version: version}
-}
-
-func TestVolume_GetClient(t *testing.T) {
-	t.Run("call on volumes with no client definitions", func(t *testing.T) {
-		// given
-		sut := Volume{
-			Name:        "data",
-			Path:        "/var/lib/scm",
-			Owner:       "",
-			Group:       "",
-			NeedsBackup: true,
-		}
-
-		// when
-		client, ok := sut.GetClient("testClient")
-
-		// then
-		assert.False(t, ok)
-		assert.Nil(t, client)
-	})
-	t.Run("call on volumes with a single client definitions", func(t *testing.T) {
-		// given
-		type testClientParams struct {
-			Type     string
-			MySecret string
-		}
-
-		sut := Volume{
-			Name:        "data",
-			Path:        "/var/lib/scm",
-			Owner:       "",
-			Group:       "",
-			NeedsBackup: true,
-			Clients: []VolumeClient{
-				{Name: "testClient", Params: testClientParams{MySecret: "supersecret", Type: "myType"}},
-			},
-		}
-
-		// when
-		client, ok := sut.GetClient("testClient")
-
-		// then
-		assert.True(t, ok)
-		require.NotNil(t, client)
-		assert.Equal(t, "testClient", client.Name)
-
-		params, ok := client.Params.(testClientParams)
-		require.True(t, ok)
-
-		assert.Equal(t, "supersecret", params.MySecret)
-		assert.Equal(t, "myType", params.Type)
-	})
-	t.Run("call on volumes with multiple client definitions", func(t *testing.T) {
-		// given
-		type testClientParams struct {
-			Type     string
-			MySecret string
-		}
-
-		sut := Volume{
-			Name:        "data",
-			Path:        "/var/lib/scm",
-			Owner:       "",
-			Group:       "",
-			NeedsBackup: true,
-			Clients: []VolumeClient{
-				{Name: "wrongClient", Params: "wrong data"},
-				{Name: "testClient", Params: testClientParams{MySecret: "supersecret", Type: "myType"}},
-			},
-		}
-
-		// when
-		client, ok := sut.GetClient("testClient")
-
-		// then
-		assert.True(t, ok)
-		require.NotNil(t, client)
-		assert.Equal(t, "testClient", client.Name)
-
-		params, ok := client.Params.(testClientParams)
-		require.True(t, ok)
-
-		assert.Equal(t, "supersecret", params.MySecret)
-		assert.Equal(t, "myType", params.Type)
-	})
-	t.Run("return false when requested client does not exits in multiple available clients", func(t *testing.T) {
-		// given
-		type testClientParams struct {
-			Type     string
-			MySecret string
-		}
-
-		sut := Volume{
-			Name:        "data",
-			Path:        "/var/lib/scm",
-			Owner:       "",
-			Group:       "",
-			NeedsBackup: true,
-			Clients: []VolumeClient{
-				{Name: "wrongClient", Params: "wrong data"},
-				{Name: "testClient", Params: testClientParams{MySecret: "supersecret", Type: "myType"}},
-			},
-		}
-
-		// when
-		client, ok := sut.GetClient("againAnotherClient")
-
-		// then
-		assert.False(t, ok)
-		require.Nil(t, client)
-	})
-}
-
-func Test_validateSecurity(t *testing.T) {
-	type args struct {
-		dogu *Dogu
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr assert.ErrorAssertionFunc
-	}{
-		{"valid empty", args{&Dogu{}}, assert.NoError},
-		{"valid add filled", args{&Dogu{Security: Security{Capabilities: Capabilities{Add: []Capability{AuditControl}}}}}, assert.NoError},
-		{"valid add filled", args{&Dogu{Security: Security{Capabilities: Capabilities{Drop: []Capability{AuditControl}}}}}, assert.NoError},
-		{"all possible values", args{&Dogu{Security: Security{Capabilities: Capabilities{Add: allCapabilities, Drop: allCapabilities}}}}, assert.NoError},
-
-		{"invalid valid add filled", args{&Dogu{Security: Security{Capabilities: Capabilities{Add: []Capability{"err"}}}}}, assert.Error},
-		{"invalid valid drop filled", args{&Dogu{Security: Security{Capabilities: Capabilities{Drop: []Capability{"err"}}}}}, assert.Error},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			tt.wantErr(t, tt.args.dogu.ValidateSecurity(), fmt.Sprintf("validateSecurity(%v)", tt.args.dogu))
-		})
-	}
-}
-
-func Test_validateSecurity_message(t *testing.T) {
-	t.Run("should match for drop errors", func(t *testing.T) {
-		// given
-		dogu := &Dogu{Name: "official/dogu", Version: "1.2.3", Security: Security{Capabilities: Capabilities{Drop: []Capability{"err"}}}}
-
-		// when
-		actual := dogu.ValidateSecurity()
-
-		// then
-		require.Error(t, actual)
-		assert.ErrorContains(t, actual, "dogu official/dogu:1.2.3 contains an invalid security field: err is not a valid capability to be dropped")
-	})
-	t.Run("should match for add errors", func(t *testing.T) {
-		// given
-		dogu := &Dogu{Name: "official/dogu", Version: "1.2.3", Security: Security{Capabilities: Capabilities{Add: []Capability{"err"}}}}
-
-		// when
-		actual := dogu.ValidateSecurity()
-
-		// then
-		require.Error(t, actual)
-		assert.ErrorContains(t, actual, "dogu official/dogu:1.2.3 contains an invalid security field: err is not a valid capability to be added")
-	})
-}
-
-func TestDogu_EffectiveCapabilities(t *testing.T) {
-	type fields struct {
-		Security Security
-	}
-	tests := []struct {
-		name   string
-		fields fields
-		want   []Capability
-	}{
-		{"drop all", fields{Security{Capabilities: Capabilities{Drop: []Capability{All}}}}, []Capability{}},
-		{"add all", fields{Security{Capabilities: Capabilities{Add: []Capability{All}}}}, allCapabilities},
-		{"drop all, add all", fields{Security{Capabilities: Capabilities{Drop: []Capability{All}, Add: []Capability{All}}}}, allCapabilities},
-		{"default list", fields{Security{Capabilities: Capabilities{}}}, DefaultCapabilities},
-		{"add 1 new and 1 existing caps to default list", fields{Security{Capabilities: Capabilities{Add: []Capability{Bpf, Chown}}}}, joinCapability(DefaultCapabilities, Bpf, Chown)},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			d := &Dogu{
-				Security: tt.fields.Security,
-			}
-			assert.ElementsMatch(t, tt.want, d.EffectiveCapabilities(), "ListCapabilities()")
-		})
-	}
-}
-
-func joinCapability(capSlice []Capability, singleCaps ...Capability) []Capability {
-	result := []Capability{}
-	result = append(result, capSlice...)
-	for _, singleCap := range singleCaps {
-		if slices.Contains(capSlice, singleCap) {
-			continue
-		}
-
-		result = append(result, singleCap)
-	}
-	return result
 }
